% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA-Combinado-DBCA.R
\name{analisis_combinado_DBCA}
\alias{analisis_combinado_DBCA}
\title{Analisis combinado de varianza para un diseño de bloques completamente aleatorizado (DBCA).}
\usage{
analisis_combinado_DBCA(
  datos,
  nombre_bloque,
  nombre_var_resp,
  nombre_tratamiento,
  nombre_entornos,
  tipo_varianza,
  efecto_bloque
)
}
\arguments{
\item{datos}{Un dataframe que contiene los datos del experimento.}

\item{nombre_bloque}{El nombre de la columna que contiene la variable de bloque.}

\item{nombre_var_resp}{El nombre de la columna que contiene la variable respuesta.}

\item{nombre_tratamiento}{El nombre de la columna que contiene la variable de tratamiento.}

\item{nombre_entornos}{El nombre de la columna que contiene la variable de entorno.}

\item{tipo_varianza}{Un entero que indica el tipo de estructura de varianza. 1 para homogeneidad y 2 para heterogeneidad.}

\item{efecto_bloque}{Un entero que indica el tipo de efecto del bloque en el modelo. 1 para aleatorio y 2 para fijo.}
}
\value{
Un dataframe con los resultados del analisis combinado de varianza.
}
\description{
Esta funcion realiza un analisis combinado de varianza para un diseño de bloques completamente aleatorizado (DBCA),
permitiendo especificar la estructura de varianza y el efecto del bloque en el modelo.
}
\examples{
# Ejemplo de uso

#data <- read.csv("datos.csv")
#resultado <- analisis_combinado_DBCA(data, "bloque", "var_resp", "tratamiento", "entorno", 1, 1)

}
