% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA-Combinado-DCA.R
\name{analisis_combinado_DCA}
\alias{analisis_combinado_DCA}
\title{Realiza análisis combinado de datos para un Diseño Completamente Aleatorizado (DCA).}
\usage{
analisis_combinado_DCA(
  datos,
  nombre_var_resp,
  nombre_tratamiento,
  nombre_entornos,
  tipo_varianza
)
}
\arguments{
\item{datos}{Un dataframe que contiene los datos del experimento.}

\item{nombre_var_resp}{El nombre de la columna que contiene la variable respuesta.}

\item{nombre_tratamiento}{El nombre de la columna que contiene los tratamientos.}

\item{nombre_entornos}{El nombre de la columna que contiene los entornos.}

\item{tipo_varianza}{Un entero que especifica el tipo de estructura de varianza a utilizar:
- 1: Sin estructura de varianza heterogénea.
- 2: Con estructura de varianza heterogénea.}
}
\value{
Devuelve el modelo ajustado e imprime en pantalla el ANOVA correspondiente a dicho modelo.
}
\description{
Esta función realiza un análisis combinado de datos para un Diseño Completamente Aleatorizado (DCA) utilizando modelos lineales mixtos.
Permite ajustar modelos con o sin estructura de varianza heterogénea entre los entornos.
}
\examples{
# Ejemplo de uso:
# analisis_combinado_DCA(datos = datos_experimento,
#                        nombre_var_resp = "variable_respuesta",
#                        nombre_tratamiento = "tratamiento",
#                        nombre_entornos = "entorno",
#                        tipo_varianza = 1)
# Este ejemplo realiza un análisis combinado sin estructura de varianza heterogénea.

}
