% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalisisCombinadoR.R
\docType{package}
\name{AnalisisCombinadoR}
\alias{AnalisisCombinadoR}
\title{AnalisisCombinadoR: Un paquete para análisis combinado de ensayos}
\description{
El paquete AnalisisCombinadoR proporciona diversas funciones para realizar
análisis combinados de ensayos. Ademas contiene funciones para realizar la
verificacion grafica de los supuestos de normalidad y de homogeneidad de
de los residuos.
}
