% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA-Exp.individuales-DBCA.R
\name{ANOVA.individuales.DBCA}
\alias{ANOVA.individuales.DBCA}
\title{Realiza análisis de varianza (ANOVA) para DBCA de forma individual para cada entorno en un conjunto de datos dado.}
\usage{
ANOVA.individuales.DBCA(
  datos,
  nombre_bloque,
  nombre_var_resp,
  nombre_tratamiento,
  nombre_entorno
)
}
\arguments{
\item{datos}{Nombre del conjunto de datos.}

\item{nombre_bloque}{Nombre de la columna donde se encuentran los bloques.}

\item{nombre_var_resp}{Nombre de la columna donde se encuentran la variable respuesta.}

\item{nombre_tratamiento}{Nombre de la columna donde se encuentran los tratamientos.}

\item{nombre_entorno}{Nombre de la columna donde se encuentran los entornos.}
}
\value{
La función devuelve una lista denominada ANOVA.ind que contiene los resultados de ANOVA y los modelos de cada entorno.
}
\description{
Realiza análisis de varianza (ANOVA) para DBCA de forma individual para cada entorno en un conjunto de datos dado.
}
\examples{
# ANOVA.individuales.DBCA(datos = datos_ejemplo,
#                         nombre_bloque = "Nombre_Bloque",
#                         nombre_var_resp = "Nombre_Var_Respuesta",
#                         nombre_tratamiento = "Nombre_Tratamiento",
#                         nombre_entorno = "Nombre_Entorno")

}
