% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA-Exp.individuales-DCA.R
\name{ANOVA.individuales.DCA}
\alias{ANOVA.individuales.DCA}
\title{#' Analisis de varianza individual para ensayos con DCA}
\usage{
ANOVA.individuales.DCA(
  datos,
  nombre_var_resp,
  nombre_tratamiento,
  nombre_entorno
)
}
\arguments{
\item{datos}{Nombre del conjunto de datos.}

\item{nombre_var_resp}{Nombre, entre comillas (""), de la columna donde se encuentran la variable respuesta.}

\item{nombre_tratamiento}{Nombre, entre comillas (""), de la columna donde se encuentran los tratamientos.}

\item{nombre_entorno}{Nombre, entre comillas (""), de la columna donde se encuentran los entornos.}
}
\value{
La funcion devuelve una lista denominada ANOVA.ind que contiene los resultados de ANOVA y los modelos de cada entorno.
}
\description{
Realiza analisis de varianza (ANOVA) para DCA de forma individual para cada entorno en un conjunto de datos.
}
\examples{
# ANOVA.individuales.DCA(datos = datos_ejemplo,
#                         nombre_var_resp = "Nombre_Var_Respuesta",
#                         nombre_tratamiento = "Nombre_Tratamiento",
#                         nombre_entorno = "Nombre_Entorno")

}
