% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graficos_residuales_homogenidad.R
\name{graficos_residuales_homogeneidad}
\alias{graficos_residuales_homogeneidad}
\title{Genera diferentes tipos de gráficos útiles para verificar el supuesto de homogeneidad de varianzas.}
\usage{
graficos_residuales_homogeneidad(ANOVA_ind, tipo_de_graficos)
}
\arguments{
\item{ANOVA_ind}{Lista de modelos ANOVA individuales.
Esta lista debe ser el resultado de la función ANOVA.individuales.DBCA o ANOVA.individuales.DCA.}

\item{tipo_de_graficos}{Tipo de gráfico que se desea generar.
Debe ser un valor entero entre 1 y 4:
- Tipo 1: Gráficos en cuadrícula de raíz cuadrada de los residuos absolutos vs. valores ajustados, cada entorno en un gráfico.
- Tipo 2: Gráficos en ventanas individuales de raíz cuadrada de los residuos absolutos vs. valores ajustados, cada entorno en un gráfico.
- Tipo 3: Gráfico de raíz cuadrada de los residuos absolutos vs. valores ajustados, todos los entornos en un gráfico diferenciando por color.
- Tipo 4: Gráfico de residuales vs. valores ajustados, todos los entornos en un gráfico diferenciando por color.
Si se proporciona un valor diferente a 1, 2, 3 o 4, la función detendrá la ejecución y devolverá un mensaje de error.}
}
\value{
La función devuelve los gráficos especificados según el tipo seleccionado.
}
\description{
Esta función genera gráficos que ayudan a evaluar si el supuesto de homogeneidad de varianzas de los residuos se cumple en los distintos entornos de un diseño experimental.
}
\examples{
# Ejemplo 1: Generar gráfico tipo 1
# graficos_residuales_homogeneidad(ANOVA_ind = resultado_ANOVA, tipo_de_graficos = 1)

# Ejemplo 2: Generar gráfico tipo 2
# graficos_residuales_homogeneidad(ANOVA_ind = resultado_ANOVA, tipo_de_graficos = 2)

}
