% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graficos.qqnorm.R
\name{graficos_residuales_qqnorm}
\alias{graficos_residuales_qqnorm}
\title{Genera graficos para verificar el supuesto de normalidad de los residuos.}
\usage{
graficos_residuales_qqnorm(ANOVA_ind, tipo_visualizacion)
}
\arguments{
\item{ANOVA_ind}{Lista de modelos ANOVA individuales.
Esta lista debe ser el resultado de la funcion ANOVA.individuales.DBCA
o ANOVA.individuales.DCA.}

\item{tipo_visualizacion}{Tipo de visualizacion de los graficos.
Debe ser un valor entero entre 1 , 2 , 3:
- 1: Graficos en cuadricula.
- 2: Graficos en ventanas individuales.
- 3: Grafico, todos los entornos en un grafico.}
}
\value{
La funcion devuelve los graficos especificados segun el tipo de
visualizacion seleccionado.
}
\description{
Esta funcion genera graficos Q-Q (quantile-quantile) para cada entorno,
permitiendo visualizar la distribucion de los residuos y compararla con
una distribucion normal teorica.
}
\examples{
# Ejemplo 1: Generar grafico tipo 1
# graficos_residuales_qqnorm(resultado_ANOVA, tipo_visualizacion = 1)

# Ejemplo 2: Generar grafico tipo 2
# graficos_residuales_qqnorm(resultado_ANOVA, tipo_visualizacion = 2)

}
