% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modificar_var_resp.R
\name{modificar_var_resp}
\alias{modificar_var_resp}
\title{Modificar la variable respuesta.}
\usage{
modificar_var_resp(
  datos,
  nombre_bloque,
  nombre_var_resp,
  nombre_tratamiento,
  nombre_entornos
)
}
\arguments{
\item{datos}{Un dataframe que contiene los datos.}

\item{nombre_bloque}{El nombre de la columna que representa los bloques en el diseño experimental.}

\item{nombre_var_resp}{El nombre de la columna que contiene la variable respuesta a ajustar.}

\item{nombre_tratamiento}{El nombre de la columna que representa los tratamientos en el diseño experimental.}

\item{nombre_entornos}{El nombre de la columna que indica los diferentes entornos en los que se realizo el experimento.}
}
\value{
Devuelve el dataframe original con una nueva columna llamada 'var_modificada', que contiene la variable respuesta ajustada por entorno.
}
\description{
Esta funcion ajusta modelos lineales para cada entorno unico en los datos,
calcula los residuos, y modifica la variable respuesta dividiendo cada valor
por la raiz cuadrada del error cuadratico medio de los residuos.
Luego, agrega esta variable respuesta modificada al conjunto de datos original.
}
\examples{
#modificar_var_resp(datos = datos_ejemplo,
#                                       nombre_bloque = "Bloque",
#                                       nombre_var_resp = "Variable_respuesta",
#                                       nombre_tratamiento = "Tratamiento",
#                                       nombre_entornos = "Entorno")
#

}
