% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graficos.qqnorm.R
\name{graficos_residuales_qqnorm}
\alias{graficos_residuales_qqnorm}
\title{Genera gráficos para verificar el supuesto de normalidad de los residuos.}
\usage{
graficos_residuales_qqnorm(ANOVA_ind, tipo_visualizacion)
}
\arguments{
\item{ANOVA_ind}{Lista de modelos ANOVA individuales.
Esta lista debe ser el resultado de la función ANOVA.individuales.DBCA
o ANOVA.individuales.DCA.}

\item{tipo_visualizacion}{Tipo de visualización de los gráficos.
Debe ser un valor entero entre 1 y 2:
- 1: Gráficos en cuadrícula.
- 2: Gráficos en ventanas individuales.
Si se proporciona un valor diferente a 1 o 2, la función detendrá
la ejecución y devolverá un mensaje de error.}
}
\value{
La función devuelve los gráficos especificados según el tipo de
visualización seleccionado.
}
\description{
Esta función genera gráficos Q-Q (quantile-quantile) para cada entorno,
permitiendo visualizar la distribución de los residuos y compararla con
una distribución normal teórica.
}
\examples{
# Ejemplo 1: Generar gráfico tipo 1
# graficos_residuales_qqnorm(resultado_ANOVA, tipo_visualizacion = 1)

# Ejemplo 2: Generar gráfico tipo 2
# graficos_residuales_qqnorm(resultado_ANOVA, tipo_visualizacion = 2)

}
